// src/front/src/app/datas/enums/pokemon-types.enum.ts
export enum PokemonType {
    Acier = 'Acier',
    Combat = 'Combat',
    Dragon = 'Dragon',
    Eau = 'Eau',
    Électrik = 'Électrik',
    Fée = 'Fée',
    Feu = 'Feu',
    Glace = 'Glace',
    Insecte = 'Insecte',
    Normal = 'Normal',
    Plante = 'Plante',
    Poison = 'Poison',
    Psy = 'Psy',
    Roche = 'Roche',
    Sol = 'Sol',
    Spectre = 'Spectre',
    Ténèbres = 'Ténèbres',
    Vol = 'Vol'
}